+++ lcode.c
@@ -762,7 +762,7 @@ static void codecomp (FuncState *fs, OpCode op, int cond, expdesc *e1,
 
 void luaK_prefix (FuncState *fs, UnOpr op, expdesc *e, int line) {
   expdesc e2;
-  e2.t = e2.f = NO_JUMP; e2.k = VKNUM; e2.u.nval = 0;
+  e2.t = e2.f = NO_JUMP; e2.k = VKNUM; e2.u.nval = static_cast<int32_t>(0);
   switch (op) {
     case OPR_MINUS: {
       if (isnumeral(e))  /* minus constant? */
