+++ lparser.c
@@ -1381,7 +1381,7 @@ static void fornum (LexState *ls, TString *varname, int line) {
   if (testnext(ls, ','))
     exp1(ls);  /* optional step */
   else {  /* default step = 1 */
-    luaK_codek(fs, fs->freereg, luaK_numberK(fs, 1));
+    luaK_codek(fs, fs->freereg, luaK_numberK(fs, static_cast<int32_t>(1)));
     luaK_reserveregs(fs, 1);
   }
   forbody(ls, base, line, 1, 1);
