+++ lua.h
@@ -18,7 +18,7 @@
 
 #define LUA_VERSION_MAJOR	"5"
 #define LUA_VERSION_MINOR	"2"
-#define LUA_VERSION_NUM		502
+#define LUA_VERSION_NUM		(static_cast<int32_t>(502))
 #define LUA_VERSION_RELEASE	"4"
 
 #define LUA_VERSION	"Lua+Eris " LUA_VERSION_MAJOR "." LUA_VERSION_MINOR
